requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
openai
pymupdf4llm
python-docx
streamlit
requests
arxiv
